more updates

+++ config.Darwin.sh
@@ -27,18 +27,18 @@ git submodule update --init
 
 # Create Links
 cd ~
-ln -s $SCRIPTDIR/bashrc .bashrc
-ln -s $SCRIPTDIR/bashrc.`uname` .bashrc.`uname`
-ln -s $SCRIPTDIR/gitconfig .gitconfig
-ln -s $SCRIPTDIR/git-global-ignore .git-global-ignore
-ln -s $SCRIPTDIR/tigrc .tigrc
-ln -s $SCRIPTDIR/vim-config .vim
-ln -s $SCRIPTDIR/slate.js .slate.js
-ln -s $SCRIPTDIR/zshrc .zshrc
-ln -s .vim/vimrc .vimrc
-ln -s .vim/gvimrc .gvimrc
+[[ ! -L .bashrc ]] && ln -s $SCRIPTDIR/bashrc .bashrc
+[[ ! -L .bashrc.`uname` ]] && ln -s $SCRIPTDIR/bashrc.`uname` .bashrc.`uname`
+[[ ! -L .gitconfig ]] && ln -s $SCRIPTDIR/gitconfig .gitconfig
+[[ ! -L .git-global-ignore ]] && ln -s $SCRIPTDIR/git-global-ignore .git-global-ignore
+[[ ! -L .tigrc ]] && ln -s $SCRIPTDIR/tigrc .tigrc
+[[ ! -L .vim ]] && ln -s $SCRIPTDIR/vim-config .vim
+[[ ! -L .slate.js ]] && ln -s $SCRIPTDIR/slate.js .slate.js
+[[ ! -L .zshrc ]] && ln -s $SCRIPTDIR/zshrc .zshrc
+[[ ! -L .vimrc ]] && ln -s .vim/vimrc .vimrc
+[[ ! -L .gvimrc ]] && ln -s .vim/gvimrc .gvimrc
 cd bin
-ln -s $SCRIPTDIR/z/z.sh
+[[ ! -L z.sh ]] && ln -s $SCRIPTDIR/z/z.sh
 
 # Copy iTerm2 Configs
 cp $SCRIPTDIR/iterm2-config/com.googlecode.iterm2.plist ~/Library/Preferences/
